fix: renamed index filename to stop Orphan script reporting falsely
